backend/requirements.txt
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<22.0
whitenoise>=6.0,<7.0
dj-database-url>=2.0,<3.0
resend>=0.5.0
requests>=2.31.0